Move Audio to new header format
