+++ CMakeLists.txt
@@ -123,6 +123,7 @@ endif()
 add_subdirectory(thirdparty)
 add_subdirectory(resources)
 add_subdirectory(graphics)
+add_subdirectory(audio)
 add_subdirectory(src)
 
 if(NOVELRT_BUILD_SAMPLES)
