+++ audio/AudioService.cpp
@@ -1,8 +1,7 @@
 // Copyright © Matt Jones and Contributors. Licensed under the MIT Licence (MIT). See LICENCE.md in the repository root
 // for more information.
 
-#include <NovelRT/Audio/Audio.h>
-#include <sndfile.h>
+#include <NovelRT/Audio/Audio.hpp>
 
 namespace NovelRT::Audio
 {
