+++ include/NovelRT/Ecs/Ecs.h
@@ -81,8 +81,8 @@ namespace NovelRT::Ecs
 #include "Catalogue.h"
 #include "SystemScheduler.h"
 #include "UnsafeComponentView.h"
-#include "Graphics/Ecs.Graphics.h"
 #include "Audio/Ecs.Audio.h"
+#include "Graphics/Ecs.Graphics.h"
 #include "Input/Ecs.Input.h"
 #include "Narrative/Ecs.Narrative.h"
 #include "Configurator.h"
